requirements.txt
# Web Framework
flask>=3.0.0
python-dotenv>=1.0.0

# Browser Automation (Anti-Detection)
camoufox[geoip]>=0.4.0
playwright>=1.40.0

# HTTP Client
requests>=2.31.0
httpx>=0.27.0

# Utilities
faker>=24.0.0

# Development
pyinstaller>=6.0.0
